% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addrowconditional.R
\name{addrowconditional}
\alias{addrowconditional}
\title{Conditionally Add Rows}
\usage{
addrowconditional(
  dataframename,
  criteriacolumn = columnname,
  repeatcount1 = 1,
  repeatcount2 = 0,
  criteria1 = filter1,
  criteria2 = NA,
  sort1 = column1,
  sort2 = NA
)
}
\arguments{
\item{dataframename}{The name of the dataframe}

\item{criteriacolumn}{The name of the column containing the criteria of interest}

\item{repeatcount1}{How many times to repeat rows meeting criteria 1. Default = 1.}

\item{repeatcount2}{How many times to repeat rows meeting criteria 1. Default = 0.}

\item{criteria1}{The conditions that meet criteria1 in the criteriacolumn}

\item{criteria2}{The conditions that meet criteria2 in the criteriacolumn}

\item{sort1}{Arrange (sort) the columns in the output based first on this value}

\item{sort2}{Arrange (sort) the columns in the output based secondly on this value}
}
\value{

}
\description{
Use this function if you need to add rows to a dataframe based on a certain criteria
For example, you might have rockfish data that had species "groups" which might
need to be partitioned into extra rows for which add'l catch can be assigned.
This function can specify which rows need to be duplicated ("criteria1"),
how many times to repeat that row ("repeatcount1"),
and then sort ("sort1") to make a more useful view.
Additionally, these are extended by secondary criteria ("criteria2"),
repeat counts ("repeatcount2"), and sorting ("sort2").
}
\examples{
addrowconditional(df, criteriacolumn = species,
                  repeatcount1 = 5, repeatcount2 = 10,
                  criteria1 = 168, criteria2 = 140,
                  sort1 = year, sort2 = species)
This would create duplicate rows 5 times for species == 168
(unspecified demersal shelf rockfishes), and 10 duplicate rows for
species == 140 (red rockfishes), then sort by columns year and species
}
