% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_pct.R
\name{count_pct}
\alias{count_pct}
\title{Summarize Proportions}
\usage{
count_pct(df)
}
\arguments{
\item{df}{The dataframe containing a group}
}
\value{

}
\description{
This function summarizes the proportions of a variable based
on the column of a grouped variables
}
\examples{
library(dplyr)
count_pct(iris \%>\%
          group_by(Species))
}
