% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicaterows.R
\name{duplicaterows}
\alias{duplicaterows}
\title{Unsummarize Data
Use this function if you encounter already summarized data and
need to split it into long format. For example, you might have
ASL data that has a row with Sex=M, Length=687, Specimen Count=3
This means that there are 3 fish that are males all with the same
length of 687. Often though, we will want to run stats on our
data which require it to NOT be summarized. This function takes
summarized data and adds rows based on the count column}
\usage{
duplicaterows(
  dataframename,
  duplicatecolname = "specimen_count",
  replacenaswithone = FALSE
)
}
\arguments{
\item{dataframename}{Name of dataframe}

\item{duplicatecolname}{Name of column that contains numeric counts of how many duplicates}

\item{replacenaswithone}{Using replacenaswithone allows you to
decide whether an NA in the count column were really a count of 1.
Carefully use this because most often an NA is a true zero.}
}
\value{

}
\description{
Unsummarize Data
Use this function if you encounter already summarized data and
need to split it into long format. For example, you might have
ASL data that has a row with Sex=M, Length=687, Specimen Count=3
This means that there are 3 fish that are males all with the same
length of 687. Often though, we will want to run stats on our
data which require it to NOT be summarized. This function takes
summarized data and adds rows based on the count column
}
\examples{
duplicaterows(dataframename = newdf, duplicatecolname = "Number_Specimens")

}
