% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_functions.R
\name{impute_cohodefault}
\alias{impute_cohodefault}
\title{Default Coho (Leon) Imputation.
This function imputes missing data (NAs) in a dataframe but does so separately
for years 1987-2000 (using a global imputation), then backwards imputes
each new year after that. This is to replicate the process used
for imputing done by Leon Shaul.
For all years pre-2000, impute globally (all values 1987-1999
inform imputed NAs). Then for each following year, calculate
a new imputation looking backwards only Leon imputed everything
once ~2000 then annually would add the new years data with blanks.
He ran the imputation on this new row which was informed by previous
years' imputations. Once imputed, these values would go in the table
and not be updated (i.e., NAs from 2002 would not be updated in
subsequent years). Make sure that all NAs are present
(a missing row is NOT same as a row with an NA).}
\usage{
impute_cohodefault(
  dfname,
  Year_column = "year",
  StreamName_column = "stream_name",
  count_column = "total_count",
  outputname = "defaultcohoimputed"
)
}
\arguments{
\item{dfname}{The name of the dataframe}

\item{Year_column}{The name of the column with the year (i.e., impute criteria 1)}

\item{StreamName_column}{The name of the column with the stream name (i.e., impute criteria 2)}

\item{count_column}{The name of the column with the counts to impute.}

\item{outputname}{Vestigial.}
}
\value{

}
\description{
Default Coho (Leon) Imputation.
This function imputes missing data (NAs) in a dataframe but does so separately
for years 1987-2000 (using a global imputation), then backwards imputes
each new year after that. This is to replicate the process used
for imputing done by Leon Shaul.
For all years pre-2000, impute globally (all values 1987-1999
inform imputed NAs). Then for each following year, calculate
a new imputation looking backwards only Leon imputed everything
once ~2000 then annually would add the new years data with blanks.
He ran the imputation on this new row which was informed by previous
years' imputations. Once imputed, these values would go in the table
and not be updated (i.e., NAs from 2002 would not be updated in
subsequent years). Make sure that all NAs are present
(a missing row is NOT same as a row with an NA).
}
\examples{
impute_cohodefault(ktn_index, Year_column="year", outputname = "ktn2020")
}
