% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_functions.R
\name{impute_local}
\alias{impute_local}
\title{10-yr Localized Imputation
This takes a dataframe with NA values and imputes missing data.
This algorithm uses "local" imputation: only 5 years before and
after impute a missing value, i.e., only using the preceding
5 years and following 5 years. Make sure that all NAs are present
(a missing row is NOT same as a row with an NA)}
\usage{
impute_local(
  dfname,
  Year_column = "year",
  StreamName_column = "stream_name",
  count_column = "total_count"
)
}
\arguments{
\item{dfname}{The name of the dataframe}

\item{Year_column}{The name of the column with the year (i.e., impute criteria 1)}

\item{StreamName_column}{The name of the column with the stream name (i.e., impute criteria 2)}

\item{count_column}{The name of the column with the counts to impute.}
}
\value{

}
\description{
10-yr Localized Imputation
This takes a dataframe with NA values and imputes missing data.
This algorithm uses "local" imputation: only 5 years before and
after impute a missing value, i.e., only using the preceding
5 years and following 5 years. Make sure that all NAs are present
(a missing row is NOT same as a row with an NA)
}
\examples{
impute_local(ktn_index, Year_column="year")
}
