% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_functions.R
\name{impute_local_improved}
\alias{impute_local_improved}
\title{10-yr Localized Imputation, improved
This takes a dataframe with NA values and imputes missing data.
This algorithm uses "local" imputation: only 5 years before and
after impute a missing value, i.e., only using the preceding
5 years and following 5 years (similar to impute_local()).
However this version adds a rule for early years (1987-1996)
to use 10 next years (10-year minimum). Make sure that all
NAs are present (a missing row is NOT same as a row with an NA)}
\usage{
impute_local_improved(
  dfname,
  Year_column = "year",
  StreamName_column = "stream_name",
  Count_column = "total_count"
)
}
\arguments{
\item{dfname}{The name of the dataframe}

\item{Year_column}{The name of the column with the year (i.e., impute criteria 1)}

\item{StreamName_column}{The name of the column with the stream name (i.e., impute criteria 2)}

\item{count_column}{The name of the column with the counts to impute.}
}
\value{

}
\description{
10-yr Localized Imputation, improved
This takes a dataframe with NA values and imputes missing data.
This algorithm uses "local" imputation: only 5 years before and
after impute a missing value, i.e., only using the preceding
5 years and following 5 years (similar to impute_local()).
However this version adds a rule for early years (1987-1996)
to use 10 next years (10-year minimum). Make sure that all
NAs are present (a missing row is NOT same as a row with an NA)
}
\examples{
impute_local_improved(ktn_index, Year_column="year")
}
