% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_crisp.R
\name{theme_crisp}
\alias{theme_crisp}
\title{Theme Crisp for ggplot2}
\usage{
theme_crisp(
  base_size = 12,
  base_family = "Arial",
  rotate_x = FALSE,
  rmborder = FALSE
)
}
\arguments{
\item{base_size}{The base font size. Default = 12.}

\item{base_family}{The base family font. Default = "Arial".}

\item{rotate_x}{TRUE/FALSE whether x axis text should be rotated 45 degrees.
Default = FALSE.}

\item{rmborder}{TRUE/FALSE whether the outer plot border should be removed.
Default = FALSE.}
}
\value{

}
\description{
Theme Crisp for ggplot2
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg)) + geom_point() + theme_crisp()
}
