% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_pct.R
\name{count_pct}
\alias{count_pct}
\title{Summarize Proportions of Grouped Variables}
\usage{
count_pct(df)
}
\arguments{
\item{df}{The dataframe containing a group}
}
\value{

}
\description{
Summarize Proportions of Grouped Variables
}
\examples{
count_pct(iris \%>\%
          group_by(Species))
}
