% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statweek.R
\name{statweek}
\alias{statweek}
\title{Calculate the Statistical Week based on a date object}
\usage{
statweek(x)
}
\arguments{
\item{x}{The date column of interest to convert}
}
\value{

}
\description{
Calculate the Statistical Week based on a date object
}
\examples{
dataframe \%>\% mutate(week = statweek(datecolumn))
}
\keyword{stat}
\keyword{week}
